ping-pong-backend/pyproject.toml
[project]
name = "ping-pong-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Van Tien NGUYEN",email = "tiennguyenhust@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.1,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "numpy (>=2.0.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
